Put Contains() in utils
