+++ utils/utils.go
@@ -17,3 +17,13 @@ type StringSorter []string
 func (a StringSorter) Len() int           { return len(a) }
 func (a StringSorter) Swap(i, j int)      { a[i], a[j] = a[j], a[i] }
 func (a StringSorter) Less(i, j int) bool { return a[i] < a[j] }
+
+// Contains tests whether string e is in slice s.
+func Contains(s []string, e string) bool {
+	for _, a := range s {
+		if a == e {
+			return true
+		}
+	}
+	return false
+}
